refactor(rendering)!: Phase 2.5 hygiene — #[must_use] transitions, rename flag_accessors, trim speculative protocol traits
Three grounded hygiene changes (audit §2.5), no behavior change.

- `#[must_use]` on the 8 self-consuming phase transitions + `run_frame`
  (`pipeline/owner/mod.rs`): `into_layout`/`into_compositing`/`into_paint`/
  `into_semantics`/`into_idle`(×3)/`finish` consume `self` and return the only
  `PipelineOwner<Phase>` handle back into the pipeline; dropping it is always a
  bug. `run_frame` uses the message form (its return is a tuple) for a better
  diagnostic.
- Rename `storage/state/flags.rs` → `storage/state/flag_accessors.rs` (git mv,
  history preserved) — the file holds inherent `RenderState` accessor impls and
  its name collided confusingly with `storage/flags.rs` (the `RenderFlags` TYPE,
  untouched). Only the `mod` declaration changed; no path references existed.
- Trim the speculative `ProtocolCompatible<Other>` + `BidirectionalProtocol`
  marker traits, the `are_protocols_compatible`/`assert_compatible` helpers,
  their marker impls, self-tests, and re-exports. Verified ZERO production
  consumers (every reference was a def/marker-impl/`#[cfg(test)]`/re-export — no
  `where`-bound, no non-test call); `Protocol`/`sealed`/erasure machinery intact.

Breaking: the `*Protocol*` compat traits leave the public surface (no consumer
named them). Gates: nextest 1527/1527 (−3 deleted speculative tests),
clippy --workspace --all-targets -D / fmt / cargo-doc -D / port-check (20+FR-033)
/ check --workspace all clean.
